requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.7
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
pydantic>=1.10.7
alembic>=1.10.2
python-jose>=3.3.0
passlib>=1.7.4
email-validator>=2.0.0